Progression on UI.
git-svn-id: http://svn.ta3d.org/yuni/trunk@1248 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/surface/cairo.h
@@ -26,7 +26,7 @@ namespace Surface
 		virtual void release();
 
 		//! Resize the surface
-		virtual void resize(unsigned int width, unsigned int height);
+		virtual void resize(float width, float height);
 
 		//! Choose a color to clear the screen with
 		virtual void clearColor(Color::RGB<uint8>& newColor);
