+++ src/yuni/gfx/window/msw/cairo.cpp
@@ -12,7 +12,8 @@ namespace Window
 	inline bool Cairo::initialize()
 	{
 		pSurface = cairo_win32_surface_create_with_ddb(GetDC(pHWND),
-			cairo_format_t.CAIRO_FORMAT_ARGB32, pWidth, pHeight);
+			cairo_format_t.CAIRO_FORMAT_ARGB32, (unsigned int)pWidth,
+			(unsigned int)pHeight);
 	}
